move snapshots into banner-store
